Move links to shared library
