+++ docs/blog/v110/index.html
@@ -798,7 +798,7 @@ code type hinting updates
 </svg>
 </span>
 logger names unification -
-<a href="../../setup/development/pub-debugger/#python-logging-for-mkdocs" id="02-setup99-development01-setting-up-debuggerpython-logging-for-mkdocs">
+<a href="../../setup/development/pub-debugger/" id="02-setup99-development01-setting-up-debugger">
 it's related to pub-debugger plugin
 </a>
 </li>
