+++ docs/setup/development/pub-debugger/index.html
@@ -1252,10 +1252,7 @@ pub-debugger
 </path>
 </svg>
 </span>
-logger names unification -
-<a href="https://mkusz.github.io/mkdocs-publisher/blog/v110/#02-setup99-development01-setting-up-debuggerpython-logging-for-mkdocs">
-it's related to pub-debugger plugin
-</a>
+logger names unification - it's related to pub-debugger plugin
 </p>
 <p></p>
 </details>
@@ -1272,10 +1269,7 @@ Changelog
 </path>
 </svg>
 </span>
-logger names unification -
-<a href="https://mkusz.github.io/mkdocs-publisher/development/changelog/#02-setup99-development01-setting-up-debuggerpython-logging-for-mkdocs">
-it's related to pub-debugger plugin
-</a>
+logger names unification - it's related to pub-debugger plugin
 </p>
 <p></p>
 </details>
