+++ docs/setup/general/pub-blog/index.html
@@ -717,7 +717,7 @@ directory, so they are not visible and do not interfere with your content.
 </p>
 <p>
 This documentation contains a
-<a href="../../../blog/">
+<a href="../../../blog/" id="blog">
 blog
 </a>
 that is created using this plugin, so you can take a look at a living example.
