+++ docs/setup/installation/index.html
@@ -654,7 +654,7 @@ There are also some external tools used, mostly by
 pub-minifier
 </code>
 plugin. Since this plugin is part of this package, but it's optional to use (not enabled by default), the process of installation of those tools is described in this
-<a href="../seo-and-sharing/pub-minifier/#tools-installation" id="03-seo-and-sharing02-setting-up-minifiertools-installation">
+<a href="../seo-and-sharing/pub-minifier/" id="03-seo-and-sharing02-setting-up-minifier">
 plugin installation
 </a>
 .
