+++ docs/setup/seo-and-sharing/pub-minifier/index.html
@@ -2058,10 +2058,10 @@ postcss-svgo - inline SVG files optimization plugin using an
 SVGO
 </a>
 tool (the same we are using for an
-<a href="./#svg-optimization">
+<a href="./" id="02-setting-up-minifier">
 SVG file optimization
 </a>
-){ #02-setting-up-minifiersvg-optimization }.
+).
 </li>
 </ul>
 <p>
