+++ docs/setup/seo-and-sharing/pub-social/index.html
@@ -709,7 +709,7 @@ Document metadata
 </h2>
 <p>
 More about document metadata you can read in the document related to setting up the
-<a href="../../general/pub-meta/#document-metadata" id="02-general01-setting-up-metadocument-metadata">
+<a href="../../general/pub-meta/" id="02-general01-setting-up-meta">
 meta plugin
 </a>
 .
