+++ docs/sitemap.xml
@@ -2,7 +2,7 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
     <url>
          <loc>https://mkusz.github.io/mkdocs-publisher/</loc>
-         <lastmod>2023-09-17</lastmod>
+         <lastmod>2023-09-26</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
@@ -107,32 +107,32 @@
     </url>
     <url>
          <loc>https://mkusz.github.io/mkdocs-publisher/blog/</loc>
-         <lastmod>2023-09-25</lastmod>
+         <lastmod>2023-09-28</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://mkusz.github.io/mkdocs-publisher/blog/1/</loc>
-         <lastmod>2023-09-25</lastmod>
+         <lastmod>2023-09-28</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://mkusz.github.io/mkdocs-publisher/blog/archive/2023/</loc>
-         <lastmod>2023-09-25</lastmod>
+         <lastmod>2023-09-28</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://mkusz.github.io/mkdocs-publisher/blog/categories/release/</loc>
-         <lastmod>2023-09-25</lastmod>
+         <lastmod>2023-09-28</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://mkusz.github.io/mkdocs-publisher/blog/tags/v0x/</loc>
-         <lastmod>2023-09-25</lastmod>
+         <lastmod>2023-09-28</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://mkusz.github.io/mkdocs-publisher/blog/tags/v1x/</loc>
-         <lastmod>2023-09-25</lastmod>
+         <lastmod>2023-09-28</lastmod>
          <changefreq>daily</changefreq>
     </url>
 </urlset>
