+++ mkdocs_publisher/obsidian/backlinks.py
@@ -23,6 +23,8 @@
 import logging
 import re
 from dataclasses import dataclass
+
+# from hashlib import md5
 from typing import Dict
 from typing import List
 from typing import Optional
@@ -31,13 +33,14 @@ from typing import cast
 from mkdocs.config.defaults import MkDocsConfig
 from mkdocs.structure.pages import Page
 
+from mkdocs_publisher._shared import links
+
 # noinspection PyProtectedMember
 from mkdocs_publisher._shared.urls import slugify
 from mkdocs_publisher.blog.plugin import BlogPlugin
 
 log = logging.getLogger("mkdocs.plugins.publisher.obsidian.backlinks")
 
-MARKDOWN_LINK_RE = re.compile(r"\[([^][\r\n]+)]\(((?!https?://)[^][)(\s]+.md)(#[\w\S]+)?\)")
 HTTP_LINK_RE = re.compile(r"\[([^][\r\n]+)]\((https?://[^][)(\s]+)(#[\w\-.]+)?\)")
 
 
@@ -49,7 +52,7 @@ class Link:
     source: str
 
 
-class Backlink:
+class BacklinkLinks:
     def __init__(
         self,
         mkdocs_config: MkDocsConfig,
@@ -63,9 +66,12 @@ class Backlink:
         """Create backlink anchor link"""
 
         anchor_link = f"{backlink}{anchor_link}"
+        # log.warning(anchor_link)
         for r in (("../", ""), (".md", ""), ("#", ""), ("/", ""), ("_", "-")):
             anchor_link = anchor_link.replace(*r)
-        return slugify(text=anchor_link)
+        anchor_link = slugify(text=anchor_link)
+        # log.error(anchor_link)
+        return anchor_link
 
     @staticmethod
     def _other_link_to_text(match: re.Match) -> str:
@@ -78,10 +84,14 @@ class Backlink:
         backlink = match.group(2)
         backlink_anchor_link = self._build_anchor_link(backlink=backlink, anchor_link=anchor_link)
         # TODO: create anchor link using slug from meta
-        return f"[{match.group(1)}]({backlink}{anchor_link}){{ #{backlink_anchor_link} }}"
+        link = f"[{match.group(1)}]({backlink}{anchor_link}){{#{backlink_anchor_link}}}"
+        # link_md5 = md5(f"[{match.group(1)}]({backlink}{anchor_link})".encode()).hexdigest()
+        # log.error(f"{link_md5} = {link}")
+        return link
 
     def _parse_markdown_link(self, match: re.Match, page: Page, line: str):
         """Parse markdown link"""
+        log.warning(match.groupdict())
         anchor_link = match.group(3) if match.group(3) is not None else ""
         original_link_destination = match.group(2)
         original_link_text = f"[{match.group(1)}]({original_link_destination}{anchor_link})"
@@ -95,7 +105,7 @@ class Backlink:
         backlink_text = line.replace(original_link_text, link_replacement)
 
         # Convert other links into text
-        backlink_text = re.sub(MARKDOWN_LINK_RE, self._other_link_to_text, backlink_text)
+        backlink_text = re.sub(links.MD_LINK_RE, self._other_link_to_text, backlink_text)
         backlink_text = re.sub(HTTP_LINK_RE, self._other_link_to_text, backlink_text)
 
         # Create a backlink with context
@@ -144,11 +154,11 @@ class Backlink:
                     self._backlinks[original_link_destination].append(link)
 
     def find_markdown_links(self, markdown: str, page: Page):
-        """Find all markdown backlinks"""
+        # """Find all markdown backlinks"""
         for line in markdown.split("\n"):
-            for match in re.finditer(MARKDOWN_LINK_RE, line):
+            for match in re.finditer(links.MD_LINK_RE, line):
                 self._parse_markdown_link(match=match, page=page, line=line)
 
     def convert_to_anchor_link(self, markdown: str) -> str:
         """Convert backlink to link with an anchor for direct navigation after clicking on it"""
-        return re.sub(MARKDOWN_LINK_RE, self._create_anchor_link, markdown)
+        return re.sub(links.MD_LINK_RE, self._create_anchor_link, markdown)
